Spool data we failed to insert to disk.
This is part of #2, but there's more to go.  In particular, I should
probably add some data for when I come around to do retries and stuff.

Also, I should add something to do retries.

There's another subtle change that slipped in here -- we're locally
computing timestamps now.  This is required to accurately replay
information.